pyproject.toml
[project]
name = "monitores-etl"
version = "0.1.0"
description = ""
authors = [
    {name = "marcosdrey",email = "programmarcos05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "scrapy (>=2.12.0,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "testcontainers (>=4.10.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'N', 'PL', 'PT', 'PD', 'AIR']

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
addopts = "-ra -q --tb=short --cov=monitores_etl"

[tool.taskipy.tasks]
lint = "ruff check"
pre_format = "ruff check --fix"
format = "ruff format"
pre_test = "task lint"
test = "pytest -s -x -vv -p no:warnings"
post_test = "coverage html"
